Add basic frame environment
This is currently just doing the 'non-fragile' grab,
as sorting out a verb. grab is a separate task.

The two optional arguments to the frame env. are collected but do nothing.

Only one slide is produced per frame, and it
is simply typeset as-is. However, this is a place
to build up other ideas.