Insert frametitle in header area
Various adjustments are needed to support this: we now tag this part
of the header and could extend to other parts of the document if required.

Likely we will need to further slit the header code to deal with the anchor, the title, etc., separately: but for the present, this works for testing.