Add support for verb material
At present this is done as an opt-in: we can then see how the
performance looks with larger documents. (I would like to always-enable,
but I see the concern re. speed.)

For the present, there is a low-level hack to allow two optional
args. for the frame env. as a result. That needs to be addressed
in ltcmd _or_ we need to have a mandatory frame title arg, which
would also deal with the issue (signature "D <> { all } m O { } c").