Split impl.of \only and onlyenv
For \only, we need to truly 'throw away' the argument: that is documented for beamer,
and needed for things that can only occur once per frame.

For onlyenv, using a b-type grabber is attractive, but fails if you then
want to collect up for example an \item in a list: the nesting of envs goes
wrong. So we have to use a box-based approach: that is now documented.