Switch to a frame/frame* approach for verbatim
Overall this feels easier to document, but does need proper testing.

Set up this way, we do no need to worry about explaining about grabbing whole frames,
but do keep the need to detail that verbatim needs special handling.

As shown in the examples, it remains easy to make a copy of verbatim-tolerant
frames.